Moving service providers to more logical place
